% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aapl_functions.R
\name{aapl_plot_prices}
\alias{aapl_plot_prices}
\title{Plot AAPL Price History}
\usage{
aapl_plot_prices(data, price_col = "close")
}
\arguments{
\item{data}{A data frame containing at least date and a price column.}

\item{price_col}{Name of the price column to plot. Defaults to "close".}
}
\value{
A ggplot object that can be printed or further customized.
}
\description{
Creates a ggplot line chart of the specified price column (default: closing
price) over time.
}
\examples{
aapl_plot_prices(aapl_sample)
}
